// lib/api/generated/models/assignmentCreateRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ServiceDesk API
 * OpenAPI spec version: 1.0
 */
import type { AssignmentCreateRequestMode } from './assignmentCreateRequestMode';

export interface AssignmentCreateRequest {
  fromLineId: number;
  mode?: AssignmentCreateRequestMode;
  /**
     * @minLength 0
     * @maxLength 1000
     */
  note?: string;
  ticketId: number;
  toLineId: number;
  toUserId?: number;
}